# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DOPIK-STOCK-MANAGEMENT API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock:
    get:
      operationId: listStock
      tags: [stock]
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockEntry"
...
  /stock/adjust:
    post:
      operationId: adjustStock
      tags: [stock]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustmentInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockAdjustment"
  /stock/adjustments:
    get:
      operationId: listStockAdjustments
      tags: [stock]
      parameters:
        - name: itemId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockAdjustment"

  # PURCHASES
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}/revert:
    post:
      operationId: revertSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  # VENDORS
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    put:
      operationId: updateVendor
      tags: [vendors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # CUSTOMERS
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUserProfile
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/password:
    put:
      operationId: updatePassword
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordUpdate"
      responses:
        "200":
          description: OK

  # DASHBOARD
...
components:
  schemas:
...
    StockEntry:
      type: object
      required: [id, itemId, itemName, qtyType, quantity, minStock, status, purchasePrice, salePrice]
      properties:
        id:
          type: integer
        itemId:
          type: integer
        itemName:
          type: string
        qtyType:
          type: string
        quantity:
          type: string
        minStock:
          type: string
        status:
          type: string
        purchasePrice:
          type: string
        salePrice:
          type: string
...
    StockAdjustmentInput:
      type: object
      required: [itemId, adjustmentType, quantity]
      properties:
        itemId:
          type: integer
        adjustmentType:
          type: string
        quantity:
          type: string
        reason:
          type: ["string", "null"]
    StockAdjustment:
      type: object
      required: [id, itemId, adjustmentType, quantity, previousQty, newQty]
      properties:
        id:
          type: integer
        itemId:
          type: integer
        itemName:
          type: ["string", "null"]
        qtyType:
          type: ["string", "null"]
        adjustmentType:
          type: string
        quantity:
          type: string
        previousQty:
          type: string
        newQty:
          type: string
        reason:
          type: ["string", "null"]
        adjustedBy:
          type: ["string", "null"]
        createdAt:
          type: string

    # PURCHASES
...
    SaleItem:
      type: object
      required: [id, saleId, itemId, quantity, unitPrice, lineTotal]
      properties:
        id:
          type: integer
        saleId:
          type: integer
        itemId:
          type: integer
        itemName:
          type: ["string", "null"]
        qtyType:
          type: ["string", "null"]
        quantity:
          type: string
        unitPrice:
          type: string
        lineTotal:
          type: string
    Sale:
      type: object
      required: [id, totalAmount, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        paymentMethod:
          type: string
        totalAmount:
          type: string
        recordedBy:
          type: ["integer", "null"]
        reverted:
          type: boolean
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
    SaleItemInput:
      type: object
      required: [itemId, quantity, unitPrice]
      properties:
        itemId:
          type: integer
        quantity:
          type: string
        unitPrice:
          type: string
    SaleInput:
      type: object
      required: [items, paymentMethod, totalAmount]
      properties:
        customerId:
          type: ["integer", "null"]
        paymentMethod:
          type: string
        totalAmount:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"

    # VENDORS
    Vendor:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string
    VendorInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        contactPerson:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]

    # CUSTOMERS
...
    UserProfile:
      type: object
      required: [id, name, email, role, status]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    PasswordUpdate:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string

    # DASHBOARD
...